requirements.txt
fastapi>=0.110
uvicorn>=0.23
jinja2>=3.1
pydantic>=2.5
python-multipart
python-docx